Refactor connect controller
- Move map related stuff into MapViewController that's used as a child controller.
- Move UI related routines from main controller into content view.
- Fix projection used to align the center of the map with activity spinner.
- Rename Connect* family of classes into Tunnel*.
- Replace delegate in (ex) ConnectViewController with a single block handler.
